Rename packages xml java and kotlin samples
