+++ samples/compose-kotlin/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="nl.dionsegijn.sample.compose">
+        package="nl.dionsegijn.xml.compose">
 
     <application
             android:allowBackup="true"
@@ -10,7 +10,7 @@
             android:supportsRtl="true"
             android:theme="@style/Theme.Konfetti">
         <activity
-                android:name="nl.dionsegijn.sample.compose.ComposeActivity"
+                android:name="nl.dionsegijn.xml.compose.ComposeActivity"
                 android:exported="true"
                 android:label="@string/app_name"
                 android:theme="@style/Theme.Konfetti.NoActionBar">
