+++ samples/compose-kotlin/src/main/java/nl/dionsegijn/xml/compose/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package nl.dionsegijn.sample.compose.ui.theme
+package nl.dionsegijn.xml.compose.ui.theme
 
 import androidx.compose.ui.graphics.Color
