+++ samples/compose-kotlin/src/main/java/nl/dionsegijn/xml/compose/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package nl.dionsegijn.sample.compose.ui.theme
+package nl.dionsegijn.xml.compose.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
