+++ samples/compose-kotlin/src/main/java/nl/dionsegijn/xml/compose/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package nl.dionsegijn.sample.compose.ui.theme
+package nl.dionsegijn.xml.compose.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
