+++ samples/xml-java/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="nl.dionsegijn.java_demo">
+        package="nl.dionsegijn.xml.java">
 
     <application
             android:allowBackup="true"
@@ -11,7 +11,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name="nl.dionsegijn.java_demo.MainActivity"
+                android:name="nl.dionsegijn.xml.java.MainActivity"
                 android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
