+++ samples/xml-java/src/main/java/nl/dionsegijn/xml/java/MainActivity.java
@@ -1,4 +1,4 @@
-package nl.dionsegijn.java_demo;
+package nl.dionsegijn.xml.java;
 
 import static nl.dionsegijn.konfetti.core.Position.Relative;
