+++ samples/xml-java/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".java.MainActivity">
 
     <nl.dionsegijn.konfetti.xml.KonfettiView
             android:id="@+id/konfettiView"
